pyproject.toml
[project]
name = "rag_082025"
version = "0.1.0"
description = "Rag Sample Demo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-chroma>=0.2.5",
    "langchain-cohere>=0.4.5",
    "langchain-community>=0.3.27",
    "langchain-core[mermaid]>=0.3.74",
    "langchain-deepinfra>=0.1.0",
    "langchain-experimental>=0.3.4",
    "langchain-neo4j>=0.3.0",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.4",
    "markdown>=3.8.2",
    "marker-pdf>=1.8.4",
    "openai>=1.99.9",
    "packaging>=24.2",
    "pinecone>=7.3.0",
    "pinecone-text>=0.11.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.48.0",
    "typing-extensions>=4.14.1",
]
